Fixed kalos bug
There was a bug that when you clicked another sort type or pokemon then clicked back to the region sort type or the back button from the pokemon entry page, the pokedex would break. It would doube some kalos entries if left on the kalos sort. This has been fixed