Fixed pokemon names, update current page num
Fixed how the pokemon names are displayed. Using the pokemon species instead of just pokemon to get their names. this makes the -normal -original tags go away. Added click functionality to the current page on the home page. Added the current page and total pages to the generation and region pages. Works correctly