Fixed current page num, last page, highlighting, page nav buttons
Fixed the current page refering to the previous page when pressing enter and changing the page number. Fixed the input for the current page on regGen. Fixed the last page issues with regGen. Fixed the next and back button issues with them disappearing. Added the highlighting feature to the current input on regGen